Split up curler.cpp
Move some functions that are related into appropriate files.
